feat(helper): isClient helper and reorganize imports (#1175)
* chore: add helper to check the browser env

* chore: use helper to check the environment

* chore: change name of environement checking mathod, change imports

* chore: fix condition

* chore: use method

* chore: change directives imports

* chore: change imports in components

* refactor: change isBrowserEnvironment to isClient

* refactor: use immedietely

* refactor: improve isClient helper
