+++ packages/vue/src/components/atoms/SfButton/SfButton.vue
@@ -11,7 +11,7 @@
   </button>
 </template>
 <script>
-import { focus } from "../../../utilities/directives/focus-directive.js";
+import { focus } from "../../../utilities/directives";
 export default {
   name: "SfButton",
   directives: {
