+++ packages/vue/src/components/atoms/SfLink/SfLink.vue
@@ -11,7 +11,7 @@
   </component>
 </template>
 <script>
-import { focus } from "../../../utilities/directives/focus-directive.js";
+import { focus } from "../../../utilities/directives";
 export default {
   name: "SfLink",
   directives: { focus },
