+++ packages/vue/src/components/molecules/SfDropdown/SfDropdown.vue
@@ -24,6 +24,8 @@
 <script>
 import SfOverlay from "../../atoms/SfOverlay/SfOverlay.vue";
 import SfButton from "../../atoms/SfButton/SfButton.vue";
+import { isClient } from "../../../utilities/helpers";
+
 export default {
   name: "SfDropdown",
   components: {
@@ -49,8 +51,7 @@ export default {
   watch: {
     isOpen: {
       handler(value) {
-        if (typeof window === "undefined" || typeof document === "undefined")
-          return;
+        if (!isClient) return;
         if (value) {
           document.addEventListener("keydown", this.keydownHandler);
         } else {
