+++ packages/vue/src/components/molecules/SfRadio/SfRadio.vue
@@ -45,7 +45,7 @@
   </div>
 </template>
 <script>
-import { focus } from "../../../utilities/directives/focus-directive.js";
+import { focus } from "../../../utilities/directives";
 export default {
   name: "SfRadio",
   directives: {
