+++ packages/vue/src/components/molecules/SfSelect/_internal/SfSelectOption.vue
@@ -14,7 +14,7 @@
   </li>
 </template>
 <script>
-import { focus } from "../../../../utilities/directives/focus-directive.js";
+import { focus } from "../../../../utilities/directives";
 export default {
   name: "SfSelectOption",
   directives: { focus },
