+++ packages/vue/src/components/organisms/SfSidebar/SfSidebar.vue
@@ -50,8 +50,9 @@
   </div>
 </template>
 <script>
-import { focusTrap } from "../../../utilities/directives/focus-trap-directive.js";
+import { focusTrap } from "../../../utilities/directives";
 import { disableBodyScroll, clearAllBodyScrollLocks } from "body-scroll-lock";
+import { isClient } from "../../../utilities/helpers";
 import SfBar from "../../molecules/SfBar/SfBar.vue";
 import SfCircleIcon from "../../atoms/SfCircleIcon/SfCircleIcon.vue";
 import SfOverlay from "../../atoms/SfOverlay/SfOverlay.vue";
@@ -115,8 +116,7 @@ export default {
   watch: {
     visible: {
       handler(value) {
-        if (typeof window === "undefined" || typeof document === "undefined")
-          return;
+        if (!isClient) return;
         if (value) {
           this.$nextTick(() => {
             disableBodyScroll(this.$refs.content);
