+++ packages/vue/src/components/organisms/SfStoreLocator/SfStoreLocator.vue
@@ -65,7 +65,7 @@
 </template>
 <script>
 import Vue from "vue";
-import { focus } from "../../../utilities/directives/focus-directive.js";
+import { focus } from "../../../utilities/directives";
 import SfIcon from "../../atoms/SfIcon/SfIcon.vue";
 import SfLoader from "../../atoms/SfLoader/SfLoader.vue";
 import SfStore from "./_internal/SfStore.vue";
