+++ packages/vue/src/components/organisms/SfTabs/_internal/SfTab.vue
@@ -25,6 +25,7 @@
 </template>
 <script>
 import { Fragment } from "vue-fragment";
+import { isClient } from "../../../../utilities/helpers";
 import SfChevron from "../../../atoms/SfChevron/SfChevron.vue";
 import SfButton from "../../../atoms/SfButton/SfButton.vue";
 export default {
@@ -51,8 +52,7 @@ export default {
   },
   methods: {
     tabClick() {
-      if (typeof window === "undefined" || typeof document === "undefined")
-        return;
+      if (!isClient) return;
       const width = Math.max(
         document.documentElement.clientWidth,
         window.innerWidth
