+++ packages/vue/src/utilities/directives/index.js
@@ -1,0 +1,2 @@
+export * from "./focus/focus-directive";
+export * from "./focus-trap/focus-trap-directive";
