+++ packages/vue/src/utilities/helpers/check-environment.js
@@ -1,0 +1,2 @@
+export const isClient = (() =>
+  typeof window !== "undefined" || typeof document !== "undefined")();
