+++ packages/vue/src/utilities/helpers/index.js
@@ -1,0 +1,2 @@
+export * from "./check-environment";
+export * from "./deprecation-warning";
